fix code expecting node style callback when argument is a callback object
These checks were written expecting a node style callback, when the callback argument in these functions is an object w/ onSuccess and onFailure props.  This pull fixes code that attempts to call this object as a function, leading to a non descriptive error being raised instead of a descriptive error being passed to the error callback.